Bind the PPPoE session to a PPP generic channel.
This seems to be necessary to receive packets. It's
also a required step on the road of getting full PPP
working, so it's not much hardship to just do it now.

As a bonus, it lets us piggyback on os.File for a bunch
of fiddly things like deadlines.

PPPoE test now passes with reading and writing of control
packets.